+++ src/commands/admin/removecourse.ts
@@ -1,5 +1,5 @@
 import { ADMIN_PERMS } from '@lib/permissions';
-import { CHANNELS, DB, SEMESTER_ID } from '@root/config';
+import { CHANNELS, DB, SEMESTER_ID } from '@root/secretVariables';
 import { Command } from '@lib/types/Command';
 import { ApplicationCommandOptionData, ApplicationCommandPermissionData, ButtonInteraction, CategoryChannel, CommandInteraction, MessageActionRow, MessageButton } from 'discord.js';
 import { modifyRoleDD } from '@root/src/lib/utils/generalUtils';
